Move home screen widget implementation to its own file
